Added experimental support for V4L2_PIX_FMT_SBGGR8 ( bayer ), V4L2_PIX_FMT_SN9C10X using code from sonix driver ( TODO : remplace bayer2rgb24 by bayer2yuv420) and V4L2_PIX_FMT_MJPEG.

+++ conf.h
@@ -71,7 +71,7 @@ struct config {
 	int control_localhost;
 	int control_html_output;
 	const char *control_authentication;
-	int frequency;
+	unsigned long frequency;
 	int tuner_number;
 	int timelapse;
 	const char *timelapse_mode;
